Move functional interfaces to a new package
This is a case of breaking backcompat to make things "n i c e r," but
eh. It's nicer. This can always be reverted if we hate it.
