+++ lib/src/main/java/org/team1540/rooster/drive/pipeline/package-info.java
@@ -1,9 +1,8 @@
 /**
  * Collection of useful inputs, processors, and outputs. Contains "stock" or "in-the-box" classes
- * that implement the {@link org.team1540.rooster.drive.pipeline.Input Input}, {@link
- * org.team1540.rooster.drive.pipeline.Processor Processor}, and {@link
- * org.team1540.rooster.drive.pipeline.Output Output} interfaces (as well as those interfaces
- * themselves).
+ * that implement the {@link org.team1540.rooster.functional.Input Input}, {@link
+ * org.team1540.rooster.functional.Processor Processor}, and {@link org.team1540.rooster.functional.Output
+ * Output} interfaces.
  *
  * <p>Unless otherwise noted, all methods, constructors, etc. in this package throw a {@link
  * java.lang.NullPointerException} if any parameters are {@code null}.
