+++ lib/src/main/java/org/team1540/rooster/functional/Executable.java
@@ -1,7 +1,7 @@
-package org.team1540.rooster.util;
+package org.team1540.rooster.functional;
 
 /**
- * Interface for actions that can be passed to a {@link SimpleCommand}.
+ * Functional interface for a function that takes no arguments and returns no output.
  */
 @FunctionalInterface
 public interface Executable {
