+++ lib/src/main/java/org/team1540/rooster/functional/Input.java
@@ -1,17 +1,15 @@
-package org.team1540.rooster.drive.pipeline;
+package org.team1540.rooster.functional;
 
 import java.util.function.Consumer;
 import java.util.function.Function;
 import java.util.function.Supplier;
-import org.team1540.rooster.util.Executable;
 
 /**
  * Extension of a {@link Supplier} adding additional composition methods.
  *
  * {@code Input} can be used exactly like {@link Supplier} (and library functions should not take
  * {@code Outputs} as method parameters, instead using {@link Supplier}). However, library functions
- * pertaining to drive pipelines should return an {@code Input} where they would normally return a
- * {@link Supplier}.
+ * should return an {@code Input} where they would normally return a {@link Supplier}.
  *
  * @param <T> The type of the input.
  */
