+++ lib/src/main/kotlin/org/team1540/rooster/drive/pipeline/PipelineExtensions.kt
@@ -2,7 +2,7 @@
 
 package org.team1540.rooster.drive.pipeline
 
-import org.team1540.rooster.util.Executable
+import org.team1540.rooster.functional.Executable
 import java.util.function.Consumer
 import java.util.function.Supplier
